backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
openai==1.6.0
anthropic==0.8.0
google-generativeai==0.3.0
langchain==0.0.350
pinecone-client==2.2.4
chromadb==0.4.20
pandas==2.1.4
numpy==1.26.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
